requirements.txt
asgiref==3.9.1
decorator==5.2.1
Django==5.2.4
djangorestframework==3.16.0
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==23.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.1
packaging==25.0
pillow==11.3.0
proglog==0.1.12
python-dotenv==1.1.1
sqlparse==0.5.3
tqdm==4.67.1
whitenoise==6.9.0
